+++ src/backend/src/main/java/RunningMachines/R2R/domain/course/entity/Course.java
@@ -1,6 +1,6 @@
 package RunningMachines.R2R.domain.course.entity;
 
-import RunningMachines.R2R.global.BaseEntity;
+import RunningMachines.R2R.global.util.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
