+++ src/backend/src/main/java/RunningMachines/R2R/global/util/BaseEntity.java
@@ -1,4 +1,4 @@
-package RunningMachines.R2R.global;
+package RunningMachines.R2R.global.util;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
